requirements.txt
asgiref==3.8.1
boto3==1.34.97
botocore==1.34.97
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.4.0
django-filter==24.2
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
googlemaps==4.10.0
idna==3.7
jmespath==1.0.1
pillow==10.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.5
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
